+++ tools/data/kinetics/rename_classnames.sh
@@ -1,7 +1,15 @@
 #!/usr/bin/env bash
 
 # Rename classname for convenience
-cd ../../../data/kinetics400/
+DATASET=$1
+if [ "$DATASET" == "kinetics400" ] || [ "$1" == "kinetics600" ] || [ "$1" == "kinetics700" ]; then
+        echo "We are processing $DATASET"
+else
+        echo "Bad Argument, we only support kinetics400, kinetics600 or kinetics700"
+        exit 0
+fi
+
+cd ../../../data/${DATASET}/
 ls ./videos_train | while read class; do \
   newclass=`echo $class | tr " " "_" | tr "(" "-" | tr ")" "-" `;
   if [ "${class}" != "${newclass}" ]
@@ -18,4 +26,4 @@ ls ./videos_val | while read class; do \
   fi
 done
 
-cd ../../tools/data/kinetics400/
+cd ../../tools/data/${DATASET}/